# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NachtragAendern": {
        "properties": {
          "betrag_netto": {
            "title": "Betrag Netto",
            "type": "integer"
          },
          "bezeichnung": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Bezeichnung",
            "type": "string"
          },
          "datum": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datum"
          },
          "stand": {
            "format": "date-time",
            "title": "Stand",
            "type": "string"
          },
          "status": {
            "default": "angeboten",
            "enum": [
              "angeboten",
              "beauftragt",
              "berechnet"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "bezeichnung",
          "betrag_netto",
          "stand"
        ],
        "title": "NachtragAendern",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nProjekt- und Finanz-Cockpit der ip³ Energietechnik GmbH.\n\nAlle Pfade liegen unter `/api`. Beträge sind ganze Cent, Zeitpunkte in UTC.\nFehlerantworten haben die Form `{code, meldung, naechster_schritt}`.\n",
    "title": "ip³ Leitstand",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/nachtraege/{nachtrag_id}": {
      "delete": {
        "description": "Angebotene und beauftragte Nachträge dürfen weg, berechnete nicht.",
        "operationId": "nachtragLoeschen",
        "parameters": [
          {
            "in": "path",
            "name": "nachtrag_id",
            "required": true,
            "schema": {
              "title": "Nachtrag Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung zahlungsplan.schreiben fehlt"
          },
          "404": {
            "description": "Projekt oder Position nicht gefunden"
          },
          "409": {
            "description": "Position gesperrt oder zwischenzeitlich geändert"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nachtrag löschen",
        "tags": [
          "Zahlungsplan"
        ]
      },
      "put": {
        "operationId": "nachtragAendern",
        "parameters": [
          {
            "in": "path",
            "name": "nachtrag_id",
            "required": true,
            "schema": {
              "title": "Nachtrag Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NachtragAendern"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NachtragAntwort"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung zahlungsplan.schreiben fehlt"
          },
          "404": {
            "description": "Projekt oder Position nicht gefunden"
          },
          "409": {
            "description": "Position gesperrt oder zwischenzeitlich geändert"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nachtrag ändern",
        "tags": [
          "Zahlungsplan"
        ]
      }
    },
...
    "/api/projekte/{projekt_nr}/nachtraege": {
      "post": {
        "operationId": "nachtragAnlegen",
        "parameters": [
          {
            "in": "path",
            "name": "projekt_nr",
            "required": true,
            "schema": {
              "title": "Projekt Nr",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NachtragEingabe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NachtragAntwort"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung zahlungsplan.schreiben fehlt"
          },
          "404": {
            "description": "Projekt oder Position nicht gefunden"
          },
          "409": {
            "description": "Position gesperrt oder zwischenzeitlich geändert"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Nachtrag anlegen",
        "tags": [
          "Zahlungsplan"
        ]
      }
    },
    "/api/projekte/{projekt_nr}/zahlungsplan": {
      "post": {
        "operationId": "zahlungsplanAnlegen",
        "parameters": [
          {
            "in": "path",
            "name": "projekt_nr",
            "required": true,
            "schema": {
              "title": "Projekt Nr",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionEingabe"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionAntwort"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung zahlungsplan.schreiben fehlt"
          },
          "404": {
            "description": "Projekt oder Position nicht gefunden"
          },
          "409": {
            "description": "Position gesperrt oder zwischenzeitlich geändert"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Zahlungsplanposition anlegen",
        "tags": [
          "Zahlungsplan"
        ]
      }
    },
...
    "/api/zahlungsplan/{position_id}": {
      "delete": {
        "description": "Löschen ist hier erlaubt – solange keine Sperre greift.\n\nEine offene Zahlungsplanposition ist eine Planung, kein Beleg: an ihr hängt nichts, und eine\nPlanung, die sich nicht korrigieren lässt, wäre in der Maske unbrauchbar. Berechnete und\nmigriert-gestellte Positionen sind dagegen Umsatz und bleiben (CLAUDE.md Regel 5).",
        "operationId": "zahlungsplanLoeschen",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "title": "Position Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung zahlungsplan.schreiben fehlt"
          },
          "404": {
            "description": "Projekt oder Position nicht gefunden"
          },
          "409": {
            "description": "Position gesperrt oder zwischenzeitlich geändert"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Zahlungsplanposition löschen",
        "tags": [
          "Zahlungsplan"
        ]
      },
      "put": {
        "operationId": "zahlungsplanAendern",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "title": "Position Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionAendern"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionAntwort"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung zahlungsplan.schreiben fehlt"
          },
          "404": {
            "description": "Projekt oder Position nicht gefunden"
          },
          "409": {
            "description": "Position gesperrt oder zwischenzeitlich geändert"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Zahlungsplanposition ändern",
        "tags": [
          "Zahlungsplan"
        ]
      }
    },
    "/api/zahlungsplan/{position_id}/gestellt-zuruecknehmen": {
      "put": {
        "description": "Eigener Weg für eine eigene Entscheidung (docs/OFFENE-PUNKTE.md Nr. 5).\n\nDer Betrag einer migriert-gestellten Position zählt zum Umsatz eines vergangenen Monats. Ihn\nim Rahmen einer gewöhnlichen Änderung mitzuverschieben würde rückwirkend Umsatz zwischen\nMonaten bewegen, ohne dass es auffällt. Deshalb: erst das Kennzeichen zurücknehmen – ein\neigener Aufruf, ein eigener Protokolleintrag –, danach ist die Position frei.",
        "operationId": "zahlungsplanGestelltZuruecknehmen",
        "parameters": [
          {
            "in": "path",
            "name": "position_id",
            "required": true,
            "schema": {
              "title": "Position Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionAntwort"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Nicht angemeldet"
          },
          "403": {
            "description": "Berechtigung zahlungsplan.schreiben fehlt"
          },
          "404": {
            "description": "Projekt oder Position nicht gefunden"
          },
          "409": {
            "description": "Position gesperrt oder zwischenzeitlich geändert"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kennzeichen „gestellt“ einer migrierten Position zurücknehmen",
        "tags": [
          "Zahlungsplan"
        ]
      }
    }
  }
}